Fix #8173: When clip's speed is close to 100%, doesn't show. The problem was that when a speed value rounded to 100, it displayed as exactly 100. I changed where the rounding occurs in the code so that when a number rounds to 100 but isn't exactly 100, it now appears as ~100%.

+++ src/projectscene/view/clipsview/cliplistitem.h
@@ -31,7 +31,7 @@ class ClipListItem : public QObject
     Q_PROPERTY(bool selected READ selected WRITE setSelected NOTIFY selectedChanged FINAL)
 
     Q_PROPERTY(int pitch READ pitch NOTIFY pitchChanged FINAL)
-    Q_PROPERTY(int speedPercentage READ speedPercentage NOTIFY speedPercentageChanged FINAL)
+    Q_PROPERTY(double speedPercentage READ speedPercentage NOTIFY speedPercentageChanged FINAL)
 
 public:
     explicit ClipListItem(QObject* parent);
@@ -66,7 +66,7 @@ public:
     void setRightVisibleMargin(double newRightVisibleMargin);
 
     int pitch() const;
-    int speedPercentage() const;
+    double speedPercentage() const;
 
 signals:
     void titleChanged();
